+++ ci/travis/build-script.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 set -ex
 
-cargo build --verbose
-cargo test --verbose
-cargo clippy --verbose -- -D warnings
-cargo doc --verbose
+cargo build --all-features
+cargo test --all-features
+cargo clippy --all-features -- -D warnings
+cargo doc --all-features
 cargo fmt --verbose -- --check