+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.4] - 2019-06-22
+### Added
+- Mock `ThreadScope` for writing tests.
+- New `with_test_support` feature to enable test utils.
+
 ## [0.1.3] - 2019-05-07
 ### Added
 - Wait for thread exit events using crossbeam-channel `Select` interface.
@@ -29,7 +34,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Threads spawning and joining.
 
 
-[Unreleased]: https://github.com/stefano-pogliani/humthreads/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/stefano-pogliani/humthreads/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/stefano-pogliani/humthreads/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/stefano-pogliani/humthreads/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/stefano-pogliani/humthreads/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/stefano-pogliani/humthreads/compare/v0.1.0...v0.1.1